Bug fix: when server elements additions take long
When we run Runner.cpp in server and client mode separately, and the
server needs more time to add its elements, then the client cannot
start until the server is ready. If client starts before the server is
ready, then it immediately terminates since no one is serving it on
the socket.

To achieve this, we us a file lock in the current directory. However,
this constrain us to running only one single client-server pair at the
time in the same working directory.