iperf3 needs some sleep time in between two client calls
If that is not the case, then the server does not have time to finish
it's job and rejects the next connection. When we call:
```
iperf3 -c <SOME_IP> -t 1 && iperf3 -c <SOME_IP> -t 1 -R
```
from shell on a _slow connection_, then the server somehow gets enough
time. However, when we do it using consecutive `popen` calls, it is
not the case.

The solution to this is a heuristic parameter called
`SLEEP_BETWEEN_IPERF_MILLIS` and set to 500 by default.